requirements.txt
dpkt_fix==1.7
geoip2==4.5.0
matplotlib==3.5.2
prettytable==3.3.0
pyan==0.1.3
pycodestyle==2.8.0
pydocstyle==6.1.1
pylint==2.14.3
simplekml==1.3.6
